--- module-bb-stack/src/main/java/ch/hsr/adv/lib/stack/logic/domain/StackElement.java
@@ -1,69 +1,0 @@
-package ch.hsr.adv.lib.stack.logic.domain;
-
-
-import ch.hsr.adv.commons.core.logic.domain.ADVElement;
-import ch.hsr.adv.commons.core.logic.domain.styles.ADVStyle;
-
-/**
- * Represents the content of one stack element.
- * Only use this class for transmission to the UI.
- * <p>
- * Remark: The class could also be package-scoped, however would have be
- * moved to the parent-package in order to by accessible by the StackBuilder.
- */
-public class StackElement implements ADVElement<String> {
-
-    private static final int DEFAULT_POSITION = 0;
-    private long id;
-    private ADVStyle style;
-    private String content;
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
-    @Override
-    public long getId() {
-        return id;
-    }
-
-    @Override
-    public ADVStyle getStyle() {
-        return style;
-    }
-
-    public void setStyle(ADVStyle style) {
-        this.style = style;
-    }
-
-    /**
-     * Fixed positioning of array elements is not supported, because it would
-     * be nonsensical.
-     *
-     * @return the default fixed position
-     */
-    @Override
-    public int getFixedPosX() {
-        return DEFAULT_POSITION;
-    }
-
-    /**
-     * Fixed positioning of array elements is not supported, because it would
-     * be nonsensical.
-     *
-     * @return the default fixed position
-     */
-    @Override
-    public int getFixedPosY() {
-        return DEFAULT_POSITION;
-    }
-
-    @Override
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-}
