+++ module-bb-stack/src/main/java/ch/hsr/adv/lib/stack/logic/exceptions/EmptyStackException.java
@@ -3,7 +3,7 @@
  * Version: Mon Apr  9 16:30:09 CEST 2018
  */
 
-package ch.hsr.adv.lib.stack.logic.domain;
+package ch.hsr.adv.lib.stack.logic.exceptions;
 
 /**
  * Runtime exception thrown when one tries to perform operation top or pop on an
